backend/requirements.txt
echo "flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
bcrypt==4.1.2" > requirements.txt